refactor(base-app): extension utils lib
