+++ apps/base-app/index.html
@@ -13,6 +13,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main/ui.ts"></script>
+    <script type="module" src="/src/app/ui/index.ts"></script>
   </body>
 </html>
