+++ apps/base-app/src/app/ui/index.ts
@@ -1,23 +1,3 @@
-import Vue from 'vue'
-import App from './ui/App.vue'
+import { mountUI } from './mount'
 
-import { withi18n } from './providers/withi18n'
-import { withRouter } from './providers/withRouter'
-import { withSanitize } from './providers/withSanitize'
-import { withVueConfig } from './providers/withVueConfig'
-
-// import './styles/theme.scss'
-import './styles/tailwind.scss'
-
-export const mountUI = async () => {
-  withVueConfig()
-  withSanitize()
-  const { i18n } = withi18n()
-  const { router } = withRouter()
-
-  new Vue({
-    i18n,
-    router,
-    render: (h) => h(App)
-  }).$mount('#app')
-}
+mountUI()
