--- apps/base-app/src/shared/lib/chrome/index.ts
@@ -1,1 +1,0 @@
-export { useChromeUtils } from './useChromeUtils'
