+++ apps/base-app/src/shared/lib/extension/extensionUtils.ts
