+++ apps/base-app/src/shared/lib/extension/extensionUtilsChrome.ts
