+++ apps/base-app/src/shared/lib/extension/extensionUtilsMock.ts
