+++ apps/base-app/src/shared/lib/extension/index.ts
@@ -1,0 +1,1 @@
+export { useExtensionUtils } from './useExtensionUtils'
